move structure queries in own namespace
