+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/services.yml
@@ -5,7 +5,7 @@ services:
     pim_catalog.event_subscriber.attribute_option_removal:
         class: 'Akeneo\Pim\Structure\Bundle\EventListener\AttributeOptionRemovalSubscriber'
         arguments:
-            - '@pim_catalog.doctrine.query.find_family_variants_identifiers_by_attribute_axes'
+            - '@akeneo.pim.structure.query.find_family_variants_identifiers_by_attribute_axes'
             - '@pim_catalog.query.product_and_product_model_query_builder_factory'
         tags:
             - { name: kernel.event_subscriber }
