+++ src/Akeneo/Pim/Structure/Bundle/Storage/Sql/FamilyVariantsByAttributeAxes.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Structure\Bundle\Doctrine\ORM\Query;
+namespace Akeneo\Pim\Structure\Bundle\Storage\Sql;
 
 use Akeneo\Pim\Structure\Component\FamilyVariant\Query\FamilyVariantsByAttributeAxesInterface;
 use Akeneo\Pim\Structure\Component\Model\FamilyVariant;
@@ -19,9 +19,6 @@ class FamilyVariantsByAttributeAxes implements FamilyVariantsByAttributeAxesInte
     /** @var EntityManagerInterface */
     private $entityManager;
 
-    /**
-     * @param EntityManagerInterface $entityManager
-     */
     public function __construct(EntityManagerInterface $entityManager)
     {
         $this->entityManager = $entityManager;
